Fix broken link to bestPlaces javascript.
Since commit 9ffa96c (Mar 25 2015), the csproj has not included the
views/bestPlaces/*.js files that are linked in the bestPlaces pages. This
was probably unnoticeable, presuming web deploy doesn't delete
non-included files from the target. However it caused a bug on new www
servers. This fix restores the correct files to the csproj <Include> list.