Some transform code for ETL. Also includes partial schema of the FDG database.
This change includes a new spec for geo data, which historically has just
been inserted into FDG database directly without evidence of any ETL. We
don't have a vendor for the geo data, so I made up a schema and filled it
with geo data from gisp-prod-db01. The schema is incomplete -- it doesn't
have a Geography field for the boundary yet.