Attempt to improve error handling related to TECH-164.
Auth failure was throwing HttpException, which is wrong and forces a 500
response. The correct exception is HttpResponseException, but even that
requires a custom error page to render it.

I couldn't make any progress using the error page system in any case. I
don't know if it's nonstandard in this project or I just am reading the
ASP.NET directions wrong. Here I resorted to returning Content object.
However, the content isn't rendered for error responses; the framework is
just returning stock error pages with no info. Hope to fix that.

On the bright side, we do return 401 instead of 500 when the key is
invalid.

We also return 400 if the key is misformatted, which previously crashed to
a 500 response. Whatever magic was converting a string to a non-nullable Guid
before calling our function does not handle failure well. So we shouldn't
put Guid objects in querystring arguments.