Implement getCustomerByKey function, per TECH-166.
I am pretty sure I didn't do the right thing with regard to Entity
Framework. I've asked for a review. The structures I defined in the
AccountController could reasonably be added to the Entity part of the
code, but they are experimental and I don't really want them in database
yet.

This commit also adds "sizeup.data" to the browser API (see TECH-164).