Remove client-side load balancing.
Originally the API client prepended 'a1-' or 'a2-' to the API hostname. I
can only imagine this was done as a round-robin load balancing scheme.
We've had server-side load balancers for years so this appears to be
totally obsolete (but inconvenient in terms of naming all our test hosts).

Removed.